requirements.txt
transformers==4.41.2
datasets
gradio
pymupdf
torch
torchvision
pillow